+++ Source/Pawnmorphs/Esoteria/Hediff_ProductionThought.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using UnityEngine;
+using RimWorld;
+using Verse;
+
+namespace Pawnmorph
+{
+    class Hediff_ProductionThought : HediffWithComps
+    {
+    }
+}
